[Feature] 목표 관련 중간 브랜치 main에 병합 #35
