+++ src/main/resources/org/iebbuda/mozi/domain/goal/mapper/GoalMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="org.iebbuda.mozi.goal.mapper.GoalMapper">
+<mapper namespace="org.iebbuda.mozi.domain.goal.mapper.GoalMapper">
 
     <!--   TABLE : UserGoal
            goal_id(pk), user_id(fk), goal_name, keyword, target_amount,
@@ -21,7 +21,7 @@
     </select>
 
 
-    <select id="get" resultType="org.iebbuda.mozi.goal.domain.GoalVO">
+    <select id="get" resultType="org.iebbuda.mozi.domain.goal.domain.GoalVO">
         select * from usergoal
         where goal_id = #{goalId}
     </select>
